requirements/local.txt
# Local development dependencies go here
-r test.txt
django-debug-toolbar==1.4
Sphinx==1.2.2
